+++ frontend/packages/data-portal/app/components/Filters/GroundTruthAnnotationFilter.tsx
@@ -4,7 +4,11 @@ import { useI18n } from 'app/hooks/useI18n'
 
 import { BooleanFilter } from './BooleanFilter'
 
-export function GroundTruthAnnotationFilter() {
+export function GroundTruthAnnotationFilter({
+  depositionPageVariant,
+}: {
+  depositionPageVariant?: boolean
+}) {
   const { t } = useI18n()
   const {
     updateValue,
@@ -12,12 +16,24 @@ export function GroundTruthAnnotationFilter() {
   } = useFilter()
 
   return (
-    <BooleanFilter
-      label={t('groundTruthAnnotation')}
-      onChange={(value) =>
-        updateValue(QueryParams.GroundTruthAnnotation, value ? 'true' : null)
-      }
-      value={isGroundTruthEnabled}
-    />
+    <>
+      <BooleanFilter
+        label={t('groundTruthAnnotation')}
+        onChange={(value) =>
+          updateValue(QueryParams.GroundTruthAnnotation, value ? 'true' : null)
+        }
+        value={isGroundTruthEnabled}
+        // FIXME: once sds upgraded to 0.20.x uncomment this
+        // caption={
+        //   depositionPageVariant ? t('depositionAnnotationsOnly') : undefined
+        // }
+      />
+      {/* FIXME: once sds upgraded to 0.20.x delete below line and remove fragment wrapper */}
+      {depositionPageVariant && (
+        <p className="pl-[32px] text-sds-body-xxs leading-sds-body-xxs text-sds-gray-500">
+          {t('depositionAnnotationsOnly')}
+        </p>
+      )}
+    </>
   )
 }