+++ frontend/packages/data-portal/app/components/Table/MetadataTable.tsx
@@ -74,23 +74,26 @@ export function MetadataTable({
                         {datum.renderValue?.(values[0]) ?? values[0]}
                       </span>
                     ))
-                    .otherwise(() => (
-                      <ul className="list-none flex flex-wrap gap-1">
-                        {values.map((value, valueIdx) => (
-                          <li
-                            className={cns(
-                              'overflow-x-auto',
-                              datum.inline && 'inline-block',
-                              datum.className,
-                            )}
-                            key={value}
-                          >
-                            {datum.renderValue?.(value) ?? value}
-                            {valueIdx < values.length - 1 && ', '}
-                          </li>
-                        ))}
-                      </ul>
-                    ))}
+                    .otherwise(
+                      () =>
+                        datum.renderValues?.(values) ?? (
+                          <ul className="list-none flex flex-wrap gap-1">
+                            {values.map((value, valueIdx) => (
+                              <li
+                                className={cns(
+                                  'overflow-x-auto',
+                                  datum.inline && 'inline-block',
+                                  datum.className,
+                                )}
+                                key={value}
+                              >
+                                {datum.renderValue?.(value) ?? value}
+                                {valueIdx < values.length - 1 && ', '}
+                              </li>
+                            ))}
+                          </ul>
+                        ),
+                    )}
                 </TableCell>
               </TableRow>
             )