+++ frontend/packages/data-portal/app/routes/depositions.$id.tsx
@@ -2,16 +2,20 @@
 
 import { CellHeaderDirection } from '@czi-sds/components'
 import { json, LoaderFunctionArgs, redirect } from '@remix-run/server-runtime'
+import { useEffect } from 'react'
 
 import { Order_By } from 'app/__generated__/graphql'
 import { apolloClient } from 'app/apollo.server'
+import { DatasetFilter } from 'app/components/DatasetFilter'
 import { DatasetsTable } from 'app/components/Deposition/DatasetsTable'
 import { DepositionHeader } from 'app/components/Deposition/DepositionHeader'
 import { TablePageLayout } from 'app/components/TablePageLayout'
 import { QueryParams } from 'app/constants/query'
+import { getDatasetsFilterData } from 'app/graphql/getDatasetsFilterData.server'
 import { getDepositionById } from 'app/graphql/getDepositionById.server'
 import { useDepositionById } from 'app/hooks/useDepositionById'
 import { useI18n } from 'app/hooks/useI18n'
+import { useDepositionHistory } from 'app/state/filterHistory'
 import { getFeatureFlag } from 'app/utils/featureFlags'
 
 export async function loader({ params, request }: LoaderFunctionArgs) {
@@ -46,27 +50,44 @@ export async function loader({ params, request }: LoaderFunctionArgs) {
     orderBy = sort === 'asc' ? Order_By.Asc : Order_By.Desc
   }
 
-  const { data } = await getDepositionById({
-    id,
-    orderBy,
-    page,
-    client: apolloClient,
-    params: url.searchParams,
-  })
+  const [depositionResponse, datasetsFilterReponse] = await Promise.all([
+    getDepositionById({
+      id,
+      orderBy,
+      page,
+      client: apolloClient,
+      params: url.searchParams,
+    }),
+    getDatasetsFilterData({
+      client: apolloClient,
+      filter: {},
+      // TODO: uncomment below when deposition fields added to backend
+      // filter: { deposition_id: { _eq: id } },
+    }),
+  ])
 
-  if (data.deposition === null) {
+  if (depositionResponse.data.deposition === null) {
     throw new Response(null, {
       status: 404,
       statusText: `Deposition with ID ${id} not found`,
     })
   }
 
-  return json(data)
+  return json({
+    depositionData: depositionResponse.data,
+    datasetsFilterData: datasetsFilterReponse.data,
+  })
 }
 
-export default function DatasetByIdPage() {
+export default function DepositionByIdPage() {
   const { deposition } = useDepositionById()
   const { t } = useI18n()
+
+  const { setPreviousDepositionId } = useDepositionHistory()
+  useEffect(
+    () => setPreviousDepositionId(deposition.id),
+    [deposition.id, setPreviousDepositionId],
+  )
 
   return (
     <TablePageLayout
@@ -78,6 +99,7 @@ export default function DatasetByIdPage() {
           totalCount: deposition.datasets_aggregate.aggregate?.count ?? 0,
           filteredCount:
             deposition.filtered_datasets_aggregate.aggregate?.count ?? 0,
+          filterPanel: <DatasetFilter depositionPageVariant />,
           countLabel: t('datasets'),
         },
       ]}